requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-multipart
google-generativeai
yt-dlp
opencv-python-headless
torch
torchvision
torchaudio
librosa
numpy
pandas
ultralytics
transformers
soundfile
openai-whisper
ffmpeg-python
mediapipe>=0.10.0
